+++ examples/Book/Action/UpdateBookAction.php
@@ -18,7 +18,7 @@ class UpdateBookAction
     public function __invoke(JsonApi $jsonApi)
     {
         // Retrieving a book domain object with an ID of $id
-        $id = $jsonApi->getRequest()->getBodyDataId();
+        $id = $jsonApi->getRequest()->getResourceId();
         $book = BookRepository::getBook($id);
         if ($book === null) {
             die("A book with an ID of '$id' can't be found!");