+++ src/JsonApi/Hydrator/UpdateHydratorTrait.php
@@ -64,7 +64,7 @@ trait UpdateHydratorTrait
         ExceptionFactoryInterface $exceptionFactory,
         $domainObject
     ) {
-        $data = $request->getBodyData();
+        $data = $request->getResource();
         if ($data === null) {
             throw $exceptionFactory->createDataMemberMissingException($request);
         }