+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGED:
 
 - ID of the hydrated resource also gets validated when it is missing
 - The provided `ExceptionFactory` can be used when validating client-generated ID-s for hydration
+- Renamed `RequestInterface::getBodyData*` methods to `RequestInterface::getResource*`
 
 REMOVED: